Move sarif-parser.ts to /common/sarif-parser.ts
